feat: refactor imports to use local_data_exporter and update model parameters for improved performance
