+++ examples/02-combined_model.py
@@ -2,7 +2,7 @@
 import os
 import pandapower as pp
 
-from data_connector import pandapower_to_dig_a_plan_schema
+from local_data_exporter import pandapower_to_dig_a_plan_schema
 from data_display.grid_plotting import plot_grid_from_pandapower
 from data_display.output_processing import compare_dig_a_plan_with_pandapower
 from pipelines import DigAPlan
@@ -42,12 +42,10 @@ base_grid_data = pandapower_to_dig_a_plan_schema(net)
 
 config = CombinedConfig(
     verbose=False,
-    big_m=1e2,
-    factor_p=1e-3,
-    factor_q=1e-3,
-    factor_v=1,
-    factor_i=1e-3,
+    big_m=1e3,
+    small_m=1,
     pipeline_type=PipelineType.COMBINED,
+    weight_infeasibility=1e-3,
 )
 dig_a_plan = DigAPlan(config=config)
