+++ src/pipelines/__init__.py
@@ -10,7 +10,9 @@ class DigAPlan:
     def __init__(self, config: PipelineConfig | BenderConfig) -> None:
 
         self.config = config or PipelineConfig()
-        self.data_manager = PipelineDataManager(self.config.big_m, self.config.small_m)
+        self.data_manager = PipelineDataManager(
+            self.config.big_m, self.config.small_m, self.config.weight_infeasibility
+        )
         if (config.pipeline_type == PipelineType.BENDER) and isinstance(
             config, BenderConfig
         ):
